requirements.txt
gdown==4.7.1
pandas==1.5.3
pydantic==1.10.7
pytorch-lightning==1.9.4
PyYAML==6.0
tqdm==4.64.1
transformers==4.24.0
wandb==0.13.10
scikit-learn==1.0.2
torch==1.13.1
